+++ crates/adapters/sp1/src/crypto.rs
@@ -17,7 +17,6 @@ pub mod private_key {
 
     use ed25519_consensus::SigningKey;
     use rand::rngs::OsRng;
-    use sov_rollup_interface::crypto::PrivateKey;
 
     use super::{SP1PublicKey, SP1Signature};
 
@@ -68,11 +67,6 @@ pub mod private_key {
         pub fn as_hex(&self) -> String {
             hex::encode(self.key_pair.to_bytes())
         }
-
-        /// Returns the address associated with the public key derived from this private key.
-        pub fn to_address<A: From<<Self as PrivateKey>::PublicKey>>(&self) -> A {
-            self.pub_key().into()
-        }
     }
 
     #[cfg(feature = "arbitrary")]
@@ -80,6 +74,7 @@ pub mod private_key {
         use proptest::prelude::{any, BoxedStrategy, Strategy};
         use rand::rngs::StdRng;
         use rand::SeedableRng;
+        use sov_rollup_interface::crypto::PrivateKey;
 
         use super::*;
 
@@ -166,11 +161,6 @@ pub struct SP1PublicKey {
 }
 
 impl SP1PublicKey {
-    /// Converts the public key to an address.
-    pub fn to_address<'a, A: From<&'a Self>>(&'a self) -> A {
-        self.into()
-    }
-
     /// Returns the bytes of the underlying public key.
     pub fn bytes(&self) -> &[u8; 32] {
         self.pub_key.as_bytes()