+++ crates/module-system/module-implementations/sov-accounts/src/capabilities.rs
@@ -7,9 +7,6 @@ impl<S: Spec> Accounts<S> {
     /// Resolve the sender's public key to an address.
     /// If the sender is not registered, but a fallback address if provided, immediately registers
     /// the credential to the fallback and then returns it.
-    ///
-    /// # Errors
-    /// If the credential is not registered AND no fallback is provided, returns an error.
     pub fn resolve_sender_address<ST: StateAccessor>(
         &self,
         default_address: &S::Address,