+++ crates/utils/transaction-generator/src/interface/mod.rs
@@ -11,7 +11,7 @@ pub use rng::*;
 use serde::{Deserialize, Serialize};
 use sov_bank::TokenId;
 use sov_modules_api::prelude::arbitrary::{self, Arbitrary};
-use sov_modules_api::{CryptoSpec, PrivateKey, Spec};
+use sov_modules_api::{CryptoSpec, PrivateKey, PublicKey, Spec};
 pub use traits::*;
 
 use super::state::ApplyToState;
@@ -407,7 +407,10 @@ where
     ) -> arbitrary::Result<(S::Address, Self::AccountView)> {
         let private_key: <<S as Spec>::CryptoSpec as CryptoSpec>::PrivateKey =
             Arbitrary::arbitrary(u)?;
-        let address: S::Address = (&private_key.pub_key()).into();
+        let address: S::Address = private_key
+            .pub_key()
+            .credential_id::<<S::CryptoSpec as CryptoSpec>::Hasher>()
+            .into();
 
         // If an account already exist with that address panic because that means our source of randomness is flawed!
         // There is only a vanishingly small change that two randomly generated addressed are the same